pyproject.toml
[project]
name = "prob-ml-pest-pipeline"
version = "0.1.0"
description = "A+ oriented synthetic pest detection pipeline for STA 561"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.2.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.4",
]

[project.scripts]
pest-pipeline = "prob_ml.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prob_ml"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]

